backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.34.0
python-dotenv==1.0.1
spotipy==2.23.0
google-generativeai==0.7.2
python-multipart==0.0.9